Rename library_h3lp_controller to conform to naming conventions
Because we call the controller the 'library-h3lp' controller, it should have an underscore in the filename too
