+++ app/javascript/controllers/index.js
@@ -58,7 +58,7 @@ application.register("gallery-preview", GalleryPreviewController)
 import GoogleCoverImageController from "./google_cover_image_controller"
 application.register("google-cover-image", GoogleCoverImageController)
 
-import LibraryH3lpController from "./libraryh3lp_controller"
+import LibraryH3lpController from "./library_h3lp_controller"
 application.register("library-h3lp", LibraryH3lpController)
 
 import LiveLookupController from "./live_lookup_controller"
