+++ app/javascript/controllers/library_h3lp_controller.js
